Test image compression
I tested out k-means image compression on an example image. This is for educational purposes, and not meant to be used as an actual image compression tool. It takes a bit too long to run as it is, and getting the best representation for a given image in a reasonable conversion time requires some tuning of the number of colors desired. This is just a proof of concept that images can be compressed using the k-means algorithm.